refactored to make overlayMessage reusable
